add proxy state machine
